update cancel modal for individual subscription
