+++ app/src/features/settings/components/BillingTeam/components/BillingDetails/modals/common/CancelPlanModal/index.scss
